refactor(py/plugins): consolidate cf and cf-ai into cloudflare-workers-ai
Merge the separate Cloudflare plugins (cf for telemetry and cf-ai for models)
into a single unified cloudflare-workers-ai plugin that provides both capabilities.

BREAKING CHANGE: No backward compatibility aliases are provided. Users must
update to use the new full names:
- CfAI → CloudflareWorkersAI
- cf_model → cloudflare_model
- cf_name → cloudflare_name
- add_cf_telemetry → add_cloudflare_telemetry
- CfConfig → CloudflareConfig
- CfEmbedConfig → CloudflareEmbedConfig

Rationale:
- Simplifies the plugin ecosystem by reducing two plugins to one
- Users get models + OTLP telemetry in a single installation
- Reduces confusion about which plugin to use for Cloudflare
- Uses full name 'cloudflare-workers-ai' for clarity
- Clean API without legacy aliases
- Marked as community plugin for clarity

Plugin consolidation:
- Renamed plugins/cf-ai to plugins/cloudflare-workers-ai
- Copied telemetry module from plugins/cf into cloudflare-workers-ai
- Deleted standalone plugins/cf directory
- Package: genkit-plugin-cloudflare-workers-ai
- Plugin name: 'cloudflare-workers-ai'
- Main class: CloudflareWorkersAI
- Helper functions: cloudflare_model, cloudflare_name
- Telemetry: add_cloudflare_telemetry
- Config: CloudflareConfig, CloudflareEmbedConfig
- Environment variables: CLOUDFLARE_ACCOUNT_ID, CLOUDFLARE_API_TOKEN

Sample consolidation:
- Merged samples/cf-ai-hello and samples/cf-hello into cloudflare-workers-ai-hello
- Updated sample to use new plugin imports

CI/CD updates:
- Updated .github/workflows/publish_python.yml to use cloudflare-workers-ai
- Cleaned up bin/fix_package_metadata.py (removed old cf entry)

Documentation updates:
- Updated py/README.md with consolidated plugin info
- Updated py/GEMINI.md package lists
- Updated plugins/README.md with community markers
- Updated samples/README.md with new sample references
- Updated all engdocs (blog, roadmap, planning docs, release guide)
- Added community plugin notices to plugin and sample READMEs
- Preserved full implementation notes with code examples

All 37 tests pass.
