+++ py/bin/fix_package_metadata.py
@@ -39,8 +39,7 @@ PLUGIN_KEYWORDS = {
     'amazon-bedrock': ['aws', 'bedrock', 'amazon'],
     'aws': ['aws', 'amazon', 'xray', 'telemetry'],
     'azure': ['azure', 'microsoft', 'application-insights', 'telemetry'],
-    'cf-ai': ['cloudflare', 'workers-ai'],
-    'cf': ['cloudflare', 'otlp', 'telemetry'],
+    'cloudflare-workers-ai': ['cloudflare', 'workers-ai', 'otlp', 'telemetry'],
     'compat-oai': ['openai', 'openai-compatible'],
     'deepseek': ['deepseek'],
     'dev-local-vectorstore': ['vector-store', 'embeddings', 'local'],
