+++ py/engdoc/planning/cloudflare-ai-plugin.md
@@ -240,7 +240,7 @@ py/plugins/cloudflare-ai/
    - Image + text inputs
 
 3. **Sample application**
-   - `cf-ai-hello` sample
+   - `cloudflare-workers-ai-hello` sample
    - README with setup instructions
 
 ## Key Implementation Details
@@ -332,7 +332,7 @@ async def _generate_stream(
 ## Sample Application
 
 ```python
-# py/samples/cf-ai-hello/src/main.py
+# py/samples/cloudflare-workers-ai-hello/src/main.py
 """Cloudflare Workers AI hello sample - Edge AI with Genkit."""
 
 from genkit.ai import Genkit
