+++ py/plugins/cloudflare-workers-ai/src/genkit/plugins/cloudflare_workers_ai/models/__init__.py
