+++ py/plugins/cloudflare-workers-ai/src/genkit/plugins/cloudflare_workers_ai/models/model_info.py
