--- py/samples/cf-hello/pyproject.toml
@@ -1,37 +1,0 @@
-# Copyright 2026 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-# SPDX-License-Identifier: Apache-2.0
-
-[project]
-authors = [{ name = "Google" }]
-dependencies = [
-  "genkit",
-  "genkit-plugin-cf",
-  "genkit-plugin-google-genai",
-  "rich>=13.0.0",
-]
-description = "Cloudflare/OTLP telemetry sample for Genkit"
-license = "Apache-2.0"
-name = "cf-hello"
-readme = "README.md"
-requires-python = ">=3.10"
-version = "0.1.0"
-
-[build-system]
-build-backend = "hatchling.build"
-requires      = ["hatchling"]
-
-[tool.hatch.build.targets.wheel]
-packages = ["src"]
