--- py/samples/cf-hello/run.sh
@@ -1,36 +1,0 @@
-#!/usr/bin/env bash
-# Copyright 2026 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-# SPDX-License-Identifier: Apache-2.0
-
-set -euo pipefail
-
-# Ensure we're in the sample directory
-cd "$(dirname "$0")"
-
-# Install dependencies if needed
-if [ ! -d ".venv" ]; then
-    uv venv
-fi
-uv sync
-
-genkit start -- \
-  uv tool run --from watchdog watchmedo auto-restart \
-    -d src \
-    -d ../../packages \
-    -d ../../plugins \
-    -p '*.py;*.prompt;*.json' \
-    -R \
-    -- uv run src/main.py "$@"
