+++ py/samples/cloudflare-workers-ai-hello/run.sh
